% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{TabCond}
\alias{TabCond}
\title{Base de dados de exemplo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38991 rows and 21 columns.
}
\usage{
TabCond
}
\description{
Descrição da base
}
\keyword{datasets}
